Move __init__ to __main__, config_parser, runner
