+++ ms2rescore/__main__.py
@@ -1,0 +1,12 @@
+"""ms2rescore."""
+
+from ms2rescore.runner import run
+
+
+def main():
+    """ms2rescore."""
+    run()
+
+
+if __name__ == "__main__":
+    main()
